Add files via upload
A program that manipulates basic string operation. It checks for trailing white spaces and an odd number of characters from an input string. If the input string has an odd number of characters then it swaps the first 2 and last 5 characters of that string. Moreover, it also prints out the capitalized version of the input string.